+++ bevy_rapier3d/examples/ray_casting3.rs
@@ -1,4 +1,5 @@
 use bevy::prelude::*;
+use bevy::window::PrimaryWindow;
 use bevy_rapier3d::prelude::*;
 
 fn main() {
@@ -72,22 +73,25 @@ pub fn setup_physics(mut commands: Commands) {
 
 fn cast_ray(
     mut commands: Commands,
-    windows: Res<Windows>,
+    windows: Query<&Window, With<PrimaryWindow>>,
     rapier_context: Res<RapierContext>,
     cameras: Query<(&Camera, &GlobalTransform)>,
 ) {
+    let window = windows.single();
+
+    let Some(cursor_position) = window.cursor_position() else { return; };
+
     // We will color in read the colliders hovered by the mouse.
-    for (camera, camera_transform) in cameras.iter() {
+    for (camera, camera_transform) in &cameras {
         // First, compute a ray from the mouse position.
-        let (ray_pos, ray_dir) =
-            ray_from_mouse_position(windows.get_primary().unwrap(), camera, camera_transform);
+        let Some(ray) = camera.viewport_to_world(camera_transform, cursor_position) else { return; };
 
         // Then cast the ray.
         let hit = rapier_context
             .cast_ray(
                 DEFAULT_WORLD_ID,
-                ray_pos,
-                ray_dir,
+                ray.origin,
+                ray.direction,
                 f32::MAX,
                 true,
                 QueryFilter::only_dynamic(),
@@ -102,26 +106,4 @@ fn cast_ray(
             commands.entity(entity).insert(ColliderDebugColor(color));
         }
     }
-}
-
-// Credit to @doomy on discord.
-fn ray_from_mouse_position(
-    window: &Window,
-    camera: &Camera,
-    camera_transform: &GlobalTransform,
-) -> (Vec3, Vec3) {
-    let mouse_position = window.cursor_position().unwrap_or(Vec2::new(0.0, 0.0));
-
-    let x = 2.0 * (mouse_position.x / window.width()) - 1.0;
-    let y = 2.0 * (mouse_position.y / window.height()) - 1.0;
-
-    let camera_inverse_matrix =
-        camera_transform.compute_matrix() * camera.projection_matrix().inverse();
-    let near = camera_inverse_matrix * Vec3::new(x, y, -1.0).extend(1.0);
-    let far = camera_inverse_matrix * Vec3::new(x, y, 1.0).extend(1.0);
-
-    let near = near.truncate() / near.w;
-    let far = far.truncate() / far.w;
-    let dir: Vec3 = far - near;
-    (near, dir)
 }