+++ src/pipeline/events.rs
@@ -45,15 +45,15 @@ pub struct ContactForceEvent {
 // investigated how to reproduce this exactly to open an
 // issue).
 /// A set of queues collecting events emitted by the physics engine.
-pub(crate) struct EventQueue<'a, 'b> {
+pub(crate) struct EventQueue<'a> {
     // Used ot retrieve the entity of colliders that have been removed from the simulation
     // since the last physics step.
     pub deleted_colliders: &'a HashMap<ColliderHandle, Entity>,
-    pub collision_events: RwLock<&'b mut EventWriter<'b, 'b, CollisionEvent>>,
-    pub contact_force_events: RwLock<&'b mut EventWriter<'b, 'b, ContactForceEvent>>,
+    pub collision_events: RwLock<EventWriter<'a, CollisionEvent>>,
+    pub contact_force_events: RwLock<EventWriter<'a, ContactForceEvent>>,
 }
 
-impl<'a, 'b> EventQueue<'a, 'b> {
+impl<'a> EventQueue<'a> {
     fn collider2entity(&self, colliders: &ColliderSet, handle: ColliderHandle) -> Entity {
         colliders
             .get(handle)
@@ -63,7 +63,7 @@ impl<'a, 'b> EventQueue<'a, 'b> {
     }
 }
 
-impl<'a, 'b> EventHandler for EventQueue<'a, 'b> {
+impl<'a> EventHandler for EventQueue<'a> {
     fn handle_collision_event(
         &self,
         _bodies: &RigidBodySet,