+++ src/plugin/plugin.rs
@@ -2,8 +2,7 @@ use crate::pipeline::{CollisionEvent, ContactForceEvent};
 use crate::plugin::configuration::SimulationToRenderTime;
 use crate::plugin::{systems, RapierConfiguration, RapierContext};
 use crate::prelude::*;
-use bevy::ecs::event::Events;
-use bevy::ecs::system::SystemParamItem;
+use bevy::ecs::{event::Events, schedule::SystemConfigs, system::SystemParamItem};
 use bevy::prelude::*;
 use std::marker::PhantomData;
 
@@ -15,9 +14,9 @@ pub use super::context::DEFAULT_WORLD_ID;
 pub type NoUserData = ();
 
 /// A plugin responsible for setting up a full Rapier physics simulation pipeline and resources.
-//
-// This will automatically setup all the resources needed to run a physics simulation with the
-// Rapier physics engine.
+///
+/// This will automatically setup all the resources needed to run a physics simulation with the
+/// Rapier physics engine.
 pub struct RapierPhysicsPlugin<PhysicsHooks = ()> {
     physics_scale: f32,
     default_system_setup: bool,
@@ -63,83 +62,70 @@ where
 
     /// Provided for use when staging systems outside of this plugin using
     /// [`with_system_setup(false)`](Self::with_system_setup).
-    /// See [`PhysicsStages`] for a description of these systems.
-    pub fn get_systems(stage: PhysicsStages) -> SystemSet {
-        match stage {
-            PhysicsStages::SyncBackend => {
-                let systems = SystemSet::new()
-                    .with_system(systems::apply_changing_worlds)
-                    .with_system(
-                        systems::update_character_controls.after(systems::apply_changing_worlds),
-                    ) // Run the character controller befor ethe manual transform propagation.
-                    .with_system(
-                        bevy::transform::transform_propagate_system
-                            .after(systems::update_character_controls),
-                    ) // Run Bevy transform propagation additionally to sync [`GlobalTransform`]
-                    .with_system(
-                        systems::init_async_colliders
-                            .after(bevy::transform::transform_propagate_system),
-                    )
-                    .with_system(systems::apply_scale.after(systems::init_async_colliders))
-                    .with_system(systems::apply_collider_user_changes.after(systems::apply_scale))
-                    .with_system(
-                        systems::apply_rigid_body_user_changes
-                            .after(systems::apply_collider_user_changes),
-                    )
-                    .with_system(
-                        systems::apply_joint_user_changes
-                            .after(systems::apply_rigid_body_user_changes),
-                    )
-                    .with_system(
-                        systems::init_rigid_bodies.after(systems::apply_joint_user_changes),
-                    )
-                    .with_system(
-                        systems::init_colliders
-                            .after(systems::init_rigid_bodies)
-                            .after(systems::init_async_colliders),
-                    )
-                    .with_system(systems::init_joints.after(systems::init_colliders))
-                    .with_system(
-                        systems::apply_initial_rigid_body_impulses.after(systems::init_colliders),
-                    )
-                    .with_system(
-                        systems::sync_removals
-                            .after(systems::init_joints)
-                            .after(systems::apply_initial_rigid_body_impulses),
-                    );
+    /// See [`PhysicsSet`] for a description of these systems.
+    pub fn get_systems(set: PhysicsSet) -> SystemConfigs {
+        // A set for `propagate_transforms` to mark it as ambiguous with `sync_simple_transforms`.
+        // Used instead of the `SystemTypeSet` as that would not allow multiple instances of the system.
+        #[derive(Debug, Hash, PartialEq, Eq, Clone, SystemSet)]
+        struct PropagateTransformsSet;
 
+        match set {
+            PhysicsSet::SyncBackend => (
+                // Change any worlds needed before doing any calculations
+                systems::apply_changing_worlds,
+                // Run the character controller before the manual transform propagation.
+                systems::update_character_controls.after(systems::apply_changing_worlds),
+                // Run Bevy transform propagation additionally to sync [`GlobalTransform`]
+                bevy::transform::systems::sync_simple_transforms
+                    .in_set(RapierTransformPropagateSet)
+                    .after(systems::update_character_controls),
+                bevy::transform::systems::propagate_transforms
+                    .after(systems::update_character_controls)
+                    .in_set(PropagateTransformsSet)
+                    .in_set(RapierTransformPropagateSet),
+                systems::init_async_colliders.after(RapierTransformPropagateSet),
+                systems::apply_scale.after(systems::init_async_colliders),
+                systems::apply_collider_user_changes.after(systems::apply_scale),
+                systems::apply_rigid_body_user_changes.after(systems::apply_collider_user_changes),
+                systems::apply_joint_user_changes.after(systems::apply_rigid_body_user_changes),
+                systems::init_rigid_bodies.after(systems::apply_joint_user_changes),
+                systems::init_colliders
+                    .after(systems::init_rigid_bodies)
+                    .after(systems::init_async_colliders),
+                systems::init_joints.after(systems::init_colliders),
+                systems::apply_initial_rigid_body_impulses.after(systems::init_colliders),
+                systems::sync_removals
+                    .after(systems::init_joints)
+                    .after(systems::apply_initial_rigid_body_impulses),
                 #[cfg(all(feature = "dim3", feature = "async-collider"))]
-                {
-                    systems.with_system(
-                        systems::init_async_scene_colliders.before(systems::init_async_colliders),
-                    )
-                }
-                #[cfg(not(feature = "dim3"))]
-                {
-                    systems
-                }
-                #[cfg(feature = "headless")]
-                {
-                    systems
-                }
-            }
-            PhysicsStages::StepSimulation => SystemSet::new()
-                .with_system(systems::step_simulation::<PhysicsHooks>)
-                .with_system(
-                    Events::<CollisionEvent>::update_system
-                        .before(systems::step_simulation::<PhysicsHooks>),
-                )
-                .with_system(
-                    Events::<ContactForceEvent>::update_system
-                        .before(systems::step_simulation::<PhysicsHooks>),
-                ),
-            PhysicsStages::Writeback => SystemSet::new()
-                .with_system(systems::update_colliding_entities)
-                .with_system(systems::writeback_rigid_bodies),
-            PhysicsStages::DetectDespawn => SystemSet::new().with_system(systems::sync_removals),
+                systems::init_async_scene_colliders
+                    .after(bevy::scene::scene_spawner_system)
+                    .before(systems::init_async_colliders),
+            )
+                .into_configs(),
+            PhysicsSet::SyncBackendFlush => (apply_system_buffers,).into_configs(),
+            PhysicsSet::StepSimulation => (
+                systems::step_simulation::<PhysicsHooks>,
+                Events::<CollisionEvent>::update_system
+                    .before(systems::step_simulation::<PhysicsHooks>),
+                Events::<ContactForceEvent>::update_system
+                    .before(systems::step_simulation::<PhysicsHooks>),
+            )
+                .into_configs(),
+            PhysicsSet::Writeback => (
+                systems::update_colliding_entities,
+                systems::writeback_rigid_bodies,
+            )
+                .into_configs(),
         }
     }
 }
+
+/// A set for rapier's copy of Bevy's transform propagation systems.
+///
+/// See [`TransformSystem`](bevy::transform::TransformSystem::TransformPropagate).
+#[derive(Debug, Hash, PartialEq, Eq, Clone, SystemSet)]
+pub struct RapierTransformPropagateSet;
 
 impl<PhysicsHooksSystemParam> Default for RapierPhysicsPlugin<PhysicsHooksSystemParam> {
     fn default() -> Self {
@@ -151,28 +137,26 @@ impl<PhysicsHooksSystemParam> Default for RapierPhysicsPlugin<PhysicsHooksSystem
     }
 }
 
-#[derive(Debug, Hash, PartialEq, Eq, Clone, StageLabel)]
 /// [`StageLabel`] for each phase of the plugin.
-pub enum PhysicsStages {
-    /// This stage runs the systems responsible for synchronizing (and
+#[derive(Debug, Hash, PartialEq, Eq, Clone, SystemSet)]
+#[system_set(base)]
+pub enum PhysicsSet {
+    /// This set runs the systems responsible for synchronizing (and
     /// initializing) backend data structures with current component state.
-    /// These systems typically run at the after [`CoreStage::Update`].
+    /// These systems typically run at the after [`CoreSet::Update`].
     SyncBackend,
+    /// The copy of [`apply_system_buffers`] that runs immediately after [`PhysicsSet::SyncBackend`].
+    SyncBackendFlush,
     /// The systems responsible for advancing the physics simulation, and
     /// updating the internal state for scene queries.
-    /// These systems typically run immediately after [`PhysicsStages::SyncBackend`].
+    /// These systems typically run immediately after [`PhysicsSet::SyncBackend`].
     StepSimulation,
     /// The systems responsible for updating
     /// [`crate::geometry::collider::CollidingEntities`] and writing
     /// the result of the last simulation step into our `bevy_rapier`
     /// components and the [`GlobalTransform`] component.
-    /// These systems typically run immediately after [`PhysicsStages::StepSimulation`].
+    /// These systems typically run immediately after [`PhysicsSet::StepSimulation`].
     Writeback,
-    /// The systems responsible for removing from Rapier the
-    /// rigid-bodies/colliders/joints which had their related `bevy_rapier`
-    /// components removed by the user (through component removal or despawn).
-    /// These systems typically run at the start of [`CoreStage::Last`].
-    DetectDespawn,
 }
 
 impl<PhysicsHooks> Plugin for RapierPhysicsPlugin<PhysicsHooks>
@@ -206,9 +190,7 @@ where
 
         // Insert all of our required resources. Don’t overwrite
         // the `RapierConfiguration` if it already exists.
-        if app.world.get_resource::<RapierConfiguration>().is_none() {
-            app.insert_resource(RapierConfiguration::default());
-        }
+        app.init_resource::<RapierConfiguration>();
 
         app.insert_resource(SimulationToRenderTime::default())
             .insert_resource(RapierContext::new(RapierWorld {
@@ -218,33 +200,33 @@ where
             .insert_resource(Events::<CollisionEvent>::default())
             .insert_resource(Events::<ContactForceEvent>::default());
 
-        // Add each stage as necessary
+        // Add each set as necessary
         if self.default_system_setup {
-            app.add_stage_after(
-                CoreStage::Update,
-                PhysicsStages::SyncBackend,
-                SystemStage::parallel()
-                    .with_system_set(Self::get_systems(PhysicsStages::SyncBackend)),
-            );
-            app.add_stage_after(
-                PhysicsStages::SyncBackend,
-                PhysicsStages::StepSimulation,
-                SystemStage::parallel()
-                    .with_system_set(Self::get_systems(PhysicsStages::StepSimulation)),
-            );
-            app.add_stage_after(
-                PhysicsStages::StepSimulation,
-                PhysicsStages::Writeback,
-                SystemStage::parallel()
-                    .with_system_set(Self::get_systems(PhysicsStages::Writeback)),
+            app.configure_sets(
+                (
+                    PhysicsSet::SyncBackend,
+                    PhysicsSet::SyncBackendFlush,
+                    PhysicsSet::StepSimulation,
+                    PhysicsSet::Writeback,
+                )
+                    .chain()
+                    .after(CoreSet::UpdateFlush)
+                    .before(CoreSet::PostUpdate),
             );
 
-            // NOTE: we run sync_removals at the end of the frame, too, in order to make sure we don’t miss any `RemovedComponents`.
-            app.add_stage_before(
-                CoreStage::Last,
-                PhysicsStages::DetectDespawn,
-                SystemStage::parallel()
-                    .with_system_set(Self::get_systems(PhysicsStages::DetectDespawn)),
+            app.add_systems(
+                Self::get_systems(PhysicsSet::SyncBackend).in_base_set(PhysicsSet::SyncBackend),
+            );
+            app.add_systems(
+                Self::get_systems(PhysicsSet::SyncBackendFlush)
+                    .in_base_set(PhysicsSet::SyncBackendFlush),
+            );
+            app.add_systems(
+                Self::get_systems(PhysicsSet::StepSimulation)
+                    .in_base_set(PhysicsSet::StepSimulation),
+            );
+            app.add_systems(
+                Self::get_systems(PhysicsSet::Writeback).in_base_set(PhysicsSet::Writeback),
             );
         }
     }